Update README.md
I added potiential objectives that we could investigate. Like there could also be any other events which could have boosted the movement of the stock such as Elon Musk's "Gamestunk" tweet.  If there are unexpected moves in the stock that we can not find correlation with the comments, we could relate them to these events. Besides, if we find the result statistically significant, we can use this anaylsis to use as trading strategies for any other stock driven by wallstreetbets community. Also we could make comparison that how reddit comments are better to show direction of stock movements that the common known trading strategies.